esoc: Call the PON and POFF hooks with bit-masked flags
Currently, the esoc driver calls the client's power-on and power-off
hooks with a boolean flag indicating if the external modem is in
error-fatal or not. However, this information itself isn't sufficient
for the clients. The clients also may rely on the actions such as, if
the driver would be shutting down the modem in the sequence or not.
Hence, convert the boolean flag to an integer based flag that carries
the bitmask of all such information.

Convert the esoc callbacks for the MHI driver as it's affected by this
API change.

Change-Id: I376e85ba499574a4c9e8cf3e845ab6311afe52d2
Signed-off-by: Raghavendra Rao Ananta <rananta@codeaurora.org>