pinctrl: msm: Remove explicit barriers from mmio ops where unneeded
For the vast majority of mmio operations in this driver, explicit memory
barriers aren't needed either because a data dependency between a read
and write already exists, or because of the presence of the spin locks
which execute a full memory barrier.

Removing all the unneeded explicit barriers considerably reduces
overhead for pinctrl operations, which in turn benefits things like i2c.

Conflicts:
	drivers/pinctrl/qcom/pinctrl-msm.c

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: Volodymyr Zhdanov <wight554@gmail.com>
Signed-off-by: pri0818 <priyanshusinghal0818@gmail.com>